rename fib to rib
