+++ go.mod
@@ -2,17 +2,18 @@ module github.com/terassyi/grp
 
 go 1.18
 
-require github.com/stretchr/testify v1.7.1
+require (
+	github.com/rs/zerolog v1.26.1
+	github.com/stretchr/testify v1.7.1
+)
 
 require (
 	github.com/inconshreveable/mousetrap v1.0.0 // indirect
-	github.com/rs/zerolog v1.26.1 // indirect
+	github.com/kr/pretty v0.1.0 // indirect
 	github.com/spf13/pflag v1.0.5 // indirect
 	github.com/vishvananda/netns v0.0.0-20191106174202-0a2b9b5464df // indirect
-	go.uber.org/atomic v1.9.0 // indirect
-	go.uber.org/multierr v1.8.0 // indirect
-	go.uber.org/zap v1.21.0 // indirect
 	golang.org/x/sys v0.0.0-20210809222454-d867a43fc93e // indirect
+	gopkg.in/check.v1 v1.0.0-20180628173108-788fd7840127 // indirect
 )
 
 require (
@@ -20,5 +21,5 @@ require (
 	github.com/pmezard/go-difflib v1.0.0 // indirect
 	github.com/spf13/cobra v1.4.0
 	github.com/vishvananda/netlink v1.1.0
-	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
+	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b
 )
